% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{DRG}
\alias{DRG}
\title{Inpatient Prospective Payment System (IPPS) for Diagnosis-Related Groups (DRG)}
\format{
A data frame with 163,065 rows and 13 variables:
}
\source{
<https://data.cms.gov/Medicare-Inpatient/Inpatient-Prospective-Payment-System-IPPS-Provider/97k6-zzx3>
}
\usage{
DRG
}
\description{
Dataset containing hospitali-specific charges for Diagnosis-Related Groups.
Pass `data("drg")`prior to using the package.
}
\keyword{datasets}
