% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drgplot.R
\name{drg_plot}
\alias{drg_plot}
\title{DRG Plot}
\usage{
drg_plot(option)
}
\arguments{
\item{option}{Options for plot variables. Includes `Average.Medicare.Payments`,
`Average.Total.Payments`, and `Average.Covered.Charges`.}
}
\value{
Returns a ggplot box-plot for payments indicated by the `option` argument.
}
\description{
This function returns a ggplot box-plot for specified payment classes.
}
\examples{
\dontrun{
data("drg")
avgmedi_boxplot <- drg_plot(option = "Average.Medicare.Payments")
}
}
