% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drgstat.R
\name{drg_stat}
\alias{drg_stat}
\title{DRG Stat}
\usage{
drg_stat(stat)
}
\arguments{
\item{stat}{Options for calculated statistic. Includes `"mean"`, `"meidan"`, and `"sd"`.}
}
\value{
Returns the statistic value for average medicare payments as per `stat`'s argument.
}
\description{
This function returns statistic values for the Average Medicare Payments
in the DRG data set.
}
\examples{
\dontrun{
drg_stat(stat = "sd")
}
}
